fix: refactor runtime
